+++ src/invent/ui/containers/page.py
@@ -20,7 +20,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from .core import Column
+from .column import Column
 
 
 class Page(Column):
