+++ src/invent/ui/core/component.py
@@ -741,169 +741,3 @@ class Container(Component):
             child.as_dict() for child in self.content
         ]
         return result
-
-
-class Column(Container):
-    """
-    A vertical container box.
-    """
-
-    @classmethod
-    def icon(cls):
-        return '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256"><path fill="currentColor" d="M104 32H64a16 16 0 0 0-16 16v160a16 16 0 0 0 16 16h40a16 16 0 0 0 16-16V48a16 16 0 0 0-16-16m0 176H64V48h40Zm88-176h-40a16 16 0 0 0-16 16v160a16 16 0 0 0 16 16h40a16 16 0 0 0 16-16V48a16 16 0 0 0-16-16m0 176h-40V48h40Z"/></svg>'  # noqa
-
-    def create_child_wrapper(self, child, index):
-        """
-        Wrap the child element in a div with grid styles set appropriately.
-        """
-        child_wrapper = document.createElement("div")
-        child_wrapper.style.setProperty("grid-column", 1)
-        child_wrapper.style.setProperty("grid-row", index)
-        child_wrapper.appendChild(child.element)
-        child.set_position(child_wrapper)
-
-        return child_wrapper
-
-    def update_child_wrapper(self, child, index):
-        """
-        Wrap the child element in a div with grid styles set appropriately.
-        """
-        child_wrapper = child.element.parentElement
-        child_wrapper.style.setProperty("grid-column", 1)
-        child_wrapper.style.setProperty("grid-row", index)
-        child_wrapper.appendChild(child.element)
-        child.set_position(child_wrapper)
-
-
-class Grid(Container):
-    """
-    A grid.
-    """
-
-    columns = IntegerProperty("Number of columns", 4)
-    column_gap = IntegerProperty("Space between columns", 0)
-    row_gap = IntegerProperty("Space between rows", 0)
-
-    @classmethod
-    def icon(cls):
-        return '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256"><path fill="currentColor" d="M216 48H40a16 16 0 0 0-16 16v128a16 16 0 0 0 16 16h176a16 16 0 0 0 16-16V64a16 16 0 0 0-16-16m-112 96v-32h48v32Zm48 16v32h-48v-32ZM40 112h48v32H40Zm64-16V64h48v32Zm64 16h48v32h-48Zm48-16h-48V64h48ZM88 64v32H40V64Zm-48 96h48v32H40Zm176 32h-48v-32h48z"/></svg>'  # noqa
-
-    def on_columns_changed(self):
-        self.element.style.gridTemplateColumns = "auto " * self.columns
-
-    def render(self):
-        """
-        Render the component.
-        """
-        element = document.createElement("div")
-        element.style.display = "grid"
-        element.style.gridTemplateColumns = "auto " * self.columns
-        element.style.columnGap = self.column_gap
-        element.style.rowGap = self.row_gap
-        element.classList.add(f"invent-{type(self).__name__.lower()}")
-
-        # Render the container's children.
-        self.render_children(element)
-
-        # Implementation detail: add child elements in the child class's own
-        # render method. See Column and Row classes for examples of this.
-        return element
-
-    def create_child_wrapper(self, child, index):
-        """
-        Wrap the child element in a div with grid styles set appropriately.
-        """
-        child_wrapper = document.createElement("div")
-
-        grid_row_span = child.row_span
-        if grid_row_span:
-            child_wrapper.style.gridRow = "span " + str(grid_row_span)
-
-        grid_column_span = child.column_span
-        if grid_column_span:
-            child_wrapper.style.gridColumn = "span " + str(grid_column_span)
-
-        child_wrapper.appendChild(child.element)
-        child.set_position(child_wrapper)
-
-        return child_wrapper
-
-    def update_child_wrapper(self, child, index):
-        """
-        Wrap the child element in a div with grid styles set appropriately.
-        """
-        child_wrapper = child.element.parentElement
-        grid_row_span = child.row_span
-        if grid_row_span:
-            child_wrapper.style.gridRow = "span " + str(grid_row_span)
-
-        grid_column_span = child.column_span
-        if grid_column_span:
-            child_wrapper.style.gridColumn = "span " + str(grid_column_span)
-
-        child_wrapper.appendChild(child.element)
-        child.set_position(child_wrapper)
-
-
-class Row(Container):
-    """
-    A horizontal container box.
-    """
-
-    @classmethod
-    def icon(cls):
-        return '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256"><path fill="currentColor" d="M208 136H48a16 16 0 0 0-16 16v40a16 16 0 0 0 16 16h160a16 16 0 0 0 16-16v-40a16 16 0 0 0-16-16m0 56H48v-40h160zm0-144H48a16 16 0 0 0-16 16v40a16 16 0 0 0 16 16h160a16 16 0 0 0 16-16V64a16 16 0 0 0-16-16m0 56H48V64h160z"/></svg>'  # noqa
-
-    def render_children(self, element):
-        """
-        Render the container's children.
-        """
-        self._update_template_columns(element)
-        super().render_children(element)
-
-    def update_children(self):
-        """
-        Update the container's children.
-        """
-        self._update_template_columns(self.element)
-        super().update_children()
-
-    def create_child_wrapper(self, child, index):
-        """
-        Wrap the child element in a div with grid styles set appropriately.
-        """
-        child_wrapper = document.createElement("div")
-        child_wrapper.style.setProperty("grid-column", index)
-        child_wrapper.style.setProperty("grid-row", 1)
-        child_wrapper.appendChild(child.element)
-        child.set_position(child_wrapper)
-
-        return child_wrapper
-
-    def update_child_wrapper(self, child, index):
-        """
-        Wrap the child element in a div with grid styles set appropriately.
-        """
-        child_wrapper = child.element.parentElement
-        child_wrapper.style.setProperty("grid-column", index)
-        child_wrapper.style.setProperty("grid-row", 1)
-        child_wrapper.appendChild(child.element)
-        child.set_position(child_wrapper)
-
-    def _update_template_columns(self, element):
-        """
-        Update the container's template columns.
-        """
-
-        template_columns = []
-        for item in self.content:
-            if (
-                item.element.classList.contains("drop-zone")
-                and len(self.content) > 1
-            ):
-                template_columns.append("0px")
-
-            else:
-                template_columns.append("auto")
-
-        element.style.gridTemplateColumns = " ".join(template_columns)
