+++ src/invent/ui/export.py
@@ -91,11 +91,15 @@ experimental_create_proxy = "auto"
 #
 "{{INVENT}}/ui/__init__.py"="{{INVENT_TO}}/ui/__init__.py"
 "{{INVENT}}/ui/app.py"="{{INVENT_TO}}/ui/app.py"
+"{{INVENT}}/ui/containers/__init__.py"="{{INVENT_TO}}/ui/containers/__init__.py"
+"{{INVENT}}/ui/containers/column.py"="{{INVENT_TO}}/ui/containers/column.py"
+"{{INVENT}}/ui/containers/grid.py"="{{INVENT_TO}}/ui/containers/grid.py"
+"{{INVENT}}/ui/containers/page.py"="{{INVENT_TO}}/ui/containers/page.py"
+"{{INVENT}}/ui/containers/row.py"="{{INVENT_TO}}/ui/containers/row.py"
 "{{INVENT}}/ui/core/__init__.py"="{{INVENT_TO}}/ui/core/__init__.py"
 "{{INVENT}}/ui/core/component.py"="{{INVENT_TO}}/ui/core/component.py"
 "{{INVENT}}/ui/core/property.py"="{{INVENT_TO}}/ui/core/property.py"
 "{{INVENT}}/ui/export.py"="{{INVENT_TO}}/ui/export.py"
-"{{INVENT}}/ui/page.py"="{{INVENT_TO}}/ui/page.py"
 "{{INVENT}}/ui/utils.py"="{{INVENT_TO}}/ui/utils.py"
 "{{INVENT}}/ui/widgets/__init__.py"="{{INVENT_TO}}/ui/widgets/__init__.py"
 "{{INVENT}}/ui/widgets/audio.py"="{{INVENT_TO}}/ui/widgets/audio.py"
